requirements.txt
fastapi==0.65.1
uvicorn==0.14.0
pymongo==4.3.2
loguru==0.6.0
aioredis==2.0.1
motor==3.1.1
python-dotenv==0.21.0

